+++ frontend/src/User/index.js
@@ -1,0 +1,6 @@
+import Login from './Login';
+import Register from './Register';
+import Forgot from './Forgot';
+import Logout from './Logout';
+
+export { Login, Logout, Register, Forgot };
